// src/vega/charts/raster/types.ts
// SPDX-FileCopyrightText: Copyright (c) 2026, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
// SPDX-License-Identifier: Apache-2.0

export type VegaMapSize = {
  width: number
  height: number
}

export type VegaMapBounds = {
  lonMin: number
  lonMax: number
  latMin: number
  latMax: number
}